# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OnlineBookshop API
  version: 1.0.0

paths:
  /books:
    get:
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/BookTitle'
        - $ref: '#/components/parameters/BookAuthor'
        - $ref: '#/components/parameters/BookGenres'
        - $ref: '#/components/parameters/BookPriceMin'
        - $ref: '#/components/parameters/BookPriceMax'
        - $ref: '#/components/parameters/BookPriceEqual'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookPage'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'

  /books/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'

  /books/{bookId}/reviews:
    get:
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookReviewRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /users:
    get:
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/UserUsername'
        - $ref: '#/components/parameters/UserEmail'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /users/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'

  /users/me:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmountUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      responses:
        '204':
          description: No Content

  /users/me/reviews:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserReviewRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/me/orders:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrentUserOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/me/orders/{orderId}:
    get:
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/me/books:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookReadResponse'

  /users/me/books/{bookId}:
    get:
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookReadResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{userId}/reviews:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserReviewRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{userId}/orders:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{userId}/orders/{orderId}:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{userId}/books:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookReadResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{userId}/books/{bookId}:
    get:
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookReadResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /reviews:
    get:
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/ReviewMinRate'
        - $ref: '#/components/parameters/ReviewMaxRate'
        - $ref: '#/components/parameters/ReviewExactRate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'

  /reviews/{reviewId}:
    get:
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'

  /orders:
    get:
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/OrderUsername'
        - $ref: '#/components/parameters/OrderMinDate'
        - $ref: '#/components/parameters/OrderMaxDate'
        - $ref: '#/components/parameters/OrderExactDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/{orderId}:
    get:
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
    Sort:
      name: sort
      in: query
      description: Repeatable Spring sort value, for example title,asc or price,desc.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    BookTitle:
      name: title
      in: query
      schema:
        type: string
    BookAuthor:
      name: author
      in: query
      schema:
        type: string
    BookGenres:
      name: genres
      in: query
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    BookPriceMin:
      name: priceMin
      in: query
      schema:
        type: number
        format: double
    BookPriceMax:
      name: priceMax
      in: query
      schema:
        type: number
        format: double
    BookPriceEqual:
      name: priceEqual
      in: query
      schema:
        type: number
        format: double
    UserUsername:
      name: username
      in: query
      schema:
        type: string
    UserEmail:
      name: email
      in: query
      schema:
        type: string
    ReviewMinRate:
      name: minRate
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
    ReviewMaxRate:
      name: maxRate
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
    ReviewExactRate:
      name: exactRate
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
    OrderUsername:
      name: username
      in: query
      schema:
        type: string
    OrderMinDate:
      name: minDate
      in: query
      schema:
        type: string
        format: date-time
    OrderMaxDate:
      name: maxDate
      in: query
      schema:
        type: string
        format: date-time
    OrderExactDate:
      name: exactDate
      in: query
      schema:
        type: string
        format: date-time

  responses:
    BadRequest:
      description: Bad Request
    NotFound:
      description: Not Found

  schemas:
    PageMetadata:
      type: object
      required:
        - currentPage
        - pageSize
        - numberOfElements
        - totalElements
        - totalPages
        - first
        - last
        - sort
      properties:
        currentPage:
          type: integer
        pageSize:
          type: integer
        numberOfElements:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        maxPageNumber:
          type:
            - integer
            - 'null'
        first:
          type: boolean
        last:
          type: boolean
        sort:
          type: array
          items:
            type: string
    BookPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Book'
    UserPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/UserResponse'
    ReviewPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Review'
    OrderPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Order'
    Book:
      type: object
      required:
        - id
        - title
        - author
        - genre
        - content
        - price
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        content:
          type: string
        price:
          type: number
          format: double
    BookRequest:
      type: object
      required:
        - title
        - author
        - genre
        - content
        - price
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        content:
          type: string
        price:
          type: number
          format: double
    BookReadResponse:
      type: object
      required:
        - id
        - title
        - author
        - genre
        - content
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        content:
          type: string
    PriceUpdateRequest:
      type: object
      required:
        - price
      properties:
        price:
          type: number
          format: double
    AmountUpdateRequest:
      type: object
      required:
        - price
      properties:
        price:
          type: number
          format: double
    UserResponse:
      type: object
      required:
        - username
        - email
        - restriction
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        restriction:
          type: boolean
    UserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
    Review:
      type: object
      required:
        - id
        - userId
        - bookId
        - rate
        - comment
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        rate:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    ReviewUpdateRequest:
      type: object
      required:
        - rate
        - comment
      properties:
        rate:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    BookReviewRequest:
      type: object
      required:
        - userId
        - rate
        - comment
      properties:
        userId:
          type: integer
          format: int64
        rate:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    UserReviewRequest:
      type: object
      required:
        - bookId
        - rate
        - comment
      properties:
        bookId:
          type: integer
          format: int64
        rate:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    Order:
      type: object
      required:
        - id
        - user
        - books
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/UserResponse'
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        createdAt:
          type: string
          format: date-time
    OrderRequest:
      type: object
      required:
        - userId
        - bookId
      properties:
        userId:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
    CurrentUserOrderRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: integer
          format: int64